+++ R/request_handler-crul.R
@@ -10,14 +10,12 @@ RequestHandlerCrul <- R6::R6Class(
     #' @return A new `RequestHandler` object
     initialize = function(request) {
       self$request_original <- request
-      self$request <- {
-        Request$new(
-          request$method,
-          request$url$url %||% request$url,
-          take_body(request),
-          request$headers
-        )
-      }
+      self$request <- vcr_request(
+        request$method,
+        request$url$url %||% request$url,
+        take_body(request),
+        request$headers
+      )
     }
   ),
   private = list(