+++ R/request_handler-httr2.R
@@ -15,14 +15,12 @@ RequestHandlerHttr2 <- R6::R6Class(
         request$method <- webmockr:::req_method_get_w(request)
       }
       self$request_original <- request
-      self$request <- {
-        Request$new(
-          request$method,
-          request$url,
-          take_body(request),
-          request$headers
-        )
-      }
+      self$request <- vcr_request(
+        request$method,
+        request$url,
+        take_body(request),
+        request$headers
+      )
     }
   ),