+++ tests/testthat/test-cassette_class.R
@@ -39,16 +39,16 @@ test_that("make_http_interaction works as expected", {
   # $response$body should be class `character`
   load("crul_resp1.rda")
   aa <- zz$make_http_interaction(crul_resp1)
-  expect_s3_class(aa$request, "Request")
-  expect_s3_class(aa$response, "VcrResponse")
+  expect_s3_class(aa$request, "vcr_request")
+  expect_s3_class(aa$response, "vcr_response")
   expect_type(aa$response$body, "character")
 
   # crul, with image response body
   # $response$body should be class `raw`
   load("crul_resp2.rda")
   bb <- zz$make_http_interaction(crul_resp2)
-  expect_s3_class(bb$request, "Request")
-  expect_s3_class(bb$response, "VcrResponse")
+  expect_s3_class(bb$request, "vcr_request")
+  expect_s3_class(bb$response, "vcr_response")
   expect_type(bb$response$body, "raw")
 })